Add files via upload
This code was used to create a new tool on ArcGIS to automate the process of editing attribute tables for parcel data in the city of Boulder, CO. It was part of a larger tool that selected data for the residents of condos to create a mailing list. This tool was used to remove records from a attribute table in ArcGIS that were redundant or were connected to polygons for parcels owned by Boulder County.